backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pydantic>=2.6.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
supabase>=2.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0
